+++ src/interfaces/cloud-foundry-options.interface.ts
@@ -1,4 +1,0 @@
-export interface ICloudFoundryOptions {
-  vcap?: {};
-  vcapFile?: string;
-}