+++ src/modules/connectors/cloud-foundry-connector/enums/cloud-foundry-service-type.enum.ts
@@ -1,0 +1,4 @@
+export enum CloudFoundryServiceType {
+  Mongodb2 = 'mongodb-2',
+  Redis2 = 'redis-2',
+}