+++ src/modules/connectors/cloud-foundry-connector/mappers/redis-credentials.mapper.ts
@@ -1,0 +1,14 @@
+import * as cfenv from 'cfenv';
+
+import { IDatabaseCredentials } from '../../../../database-metrics-logger';
+import { CloudFoundryServiceType } from '../enums';
+import { serviceTypeDatabaseTypeMapper } from './service-types.mapper';
+
+export const mapRedisCredentials = (cloudFoundryCredentials: cfenv.IRedisCredentials): IDatabaseCredentials => {
+  return {
+    databaseType: serviceTypeDatabaseTypeMapper.get(CloudFoundryServiceType.Redis2),
+    host: cloudFoundryCredentials.host,
+    port: cloudFoundryCredentials.port,
+    password: cloudFoundryCredentials.password,
+  };
+};