+++ test/config/teardown.ts
@@ -1,0 +1,3 @@
+export default async (): Promise<void> => {
+  await (global as any).__MONGOD__.stop();
+};